fixed verify feature

+++ src/main/java/com/springboot/minimartapi/role/RoleRepo.java
@@ -1,9 +1,13 @@
 package com.springboot.minimartapi.role;
 
+import com.springboot.minimartapi.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepo extends JpaRepository<Role, Integer> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
+
+
 
 }
